move utils to common
